requirements.txt
pandas
numpy
requests
shapely
pyproj
scikit-learn
openpyxl
pytz
python-dotenv
